chore: move user kx utils to users.ts
