+++ frontend/utils/crypto/general.ts
@@ -288,24 +288,6 @@ export const decryptAsymmetric = async (
   })
 }
 
-export const getUserKxPublicKey = async (signingPublicKey: string) => {
-  await _sodium.ready
-  const sodium = _sodium
-
-  return sodium.to_hex(
-    sodium.crypto_sign_ed25519_pk_to_curve25519(sodium.from_hex(signingPublicKey))
-  )
-}
-
-export const getUserKxPrivateKey = async (signingPrivateKey: string) => {
-  await _sodium.ready
-  const sodium = _sodium
-
-  return sodium.to_hex(
-    sodium.crypto_sign_ed25519_sk_to_curve25519(sodium.from_hex(signingPrivateKey))
-  )
-}
-
 export const digest = async (input: string, salt: string) => {
   await _sodium.ready
   const sodium = _sodium
